requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
razorpay==2.0.1
python-dotenv==1.2.2
gpiozero
pigpio
